feat(ips): move mocks into src/__mocks__
ref: #MANAGER-18212
