+++ packages/manager/apps/ips/src/__mocks__/catalog/catalog.ts
