+++ packages/manager/apps/ips/src/__mocks__/catalog/dedicated-server-catalog.ts
