+++ packages/manager/apps/ips/src/__mocks__/catalog/pcc-catalog.ts
@@ -1,5 +1,5 @@
 import { CurrencyCode } from '@ovh-ux/manager-react-components';
-import { PccCatalogResponse } from '../../src/data/api';
+import { PccCatalogResponse } from '@/data/api';
 
 export const catalogPcc: PccCatalogResponse = [
   {
