+++ packages/manager/apps/ips/src/__mocks__/index.ts
