+++ packages/manager/apps/ips/src/__mocks__/ip/get-ip-details.json
