+++ packages/manager/apps/ips/src/pages/listing/ipListing/components/DatagridCells/IpAlerts/IpAlerts.spec.tsx
@@ -3,7 +3,7 @@ import { describe, it, vi } from 'vitest';
 import { render } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ListingContext } from '@/pages/listing/listingContext';
-import ipDetailsList from '../../../../../../../mocks/ip/get-ip-details.json';
+import ipDetailsList from '@/__mocks__/ip/get-ip-details.json';
 import { IpAlerts, IpAlertsProps } from './IpAlerts';
 import {
   IpAntihackType,
