+++ packages/manager/apps/ips/src/pages/listing/ipListing/components/DatagridCells/IpCell/IpCell.spec.tsx
@@ -4,7 +4,7 @@ import { describe, expect, it, vi } from 'vitest';
 import { render, waitFor } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { IpCell, IpCellProps } from './IpCell';
-import ipDetailsList from '../../../../../../../mocks/ip/get-ip-details.json';
+import ipDetailsList from '@/__mocks__/ip/get-ip-details.json';
 
 const queryClient = new QueryClient();
 /** MOCKS */
