+++ packages/manager/apps/ips/src/pages/listing/manageOrganisations/manage.organisations.spec.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { describe, it, expect, vi } from 'vitest';
 import { render, waitFor } from '@testing-library/react';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import { organisationMockDetails } from '../../../../mocks';
+import { organisationMockDetails } from '@/__mocks__';
 import { ManageOrganisationsDatagrid } from './components/OrganisationsDatagrid/manageOrganisationsDataGrid.component';
 
 const queryClient = new QueryClient();
